requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
ultralytics>=8.0.0
torch>=2.0.0
pandas>=2.0.0
Pillow>=10.0.0
flask>=2.0.0
werkzeug>=2.0.0